refactor(Exception): 예외처리 부분 다른 패키지로 변경
