+++ src/main/java/racingcar/view/InputView.java
@@ -1,5 +1,9 @@
 package racingcar.view;
 
+import static racingcar.view.Exception.validateAttempts;
+import static racingcar.view.Exception.validateCarCount;
+import static racingcar.view.Exception.validateNameLength;
+
 import camp.nextstep.edu.missionutils.Console;
 import java.util.ArrayList;
 import java.util.List;
@@ -31,24 +35,6 @@ public class InputView {
             return Integer.parseInt(num);
         } catch (IllegalArgumentException e) {
             throw new IllegalArgumentException("입력값은 정수여야합니다.");
-        }
-    }
-
-    public static void validateAttempts(int attempts) {
-        if (attempts < 1) {
-            throw new IllegalArgumentException("시도횟수는 1보다 커야합니다.");
-        }
-    }
-
-    public static void validateCarCount(int carCount) {
-        if (carCount < 2) {
-            throw new IllegalArgumentException("참여 자동차는 2대 이상이어야 합니다.");
-        }
-    }
-
-    public static void validateNameLength(String carName) {
-        if (carName.length() > 5) {
-            throw new IllegalArgumentException("자동차 이름은 5자 이하여야 합니다.");
         }
     }
 }
